+++ server.go
@@ -89,7 +89,7 @@ func NewWSServer(opts ...ServerOption) Server {
 func NewWSSServer(opts ...ServerOption) Server {
 	s := newServer(WSS_SERVER, opts...)
 
-	if s.addr == "" || s.cert == "" || s.privateKey == "" || s.caCert == "" {
+	if s.addr == "" || s.cert == "" || s.privateKey == "" {
 		panic(fmt.Sprintf("@addr:%s, @cert:%s, @privateKey:%s, @caCert:%s",
 			s.addr, s.cert, s.privateKey, s.caCert))
 	}
@@ -237,7 +237,8 @@ func (s *server) runTcpEventLoop(newSession NewSessionCallback) {
 				return
 			}
 			if delay != 0 {
-				time.Sleep(delay)
+				// time.Sleep(delay)
+				<-wheel.After(delay)
 			}
 			client, err = s.accept(newSession)
 			if err != nil {